Initial commit
Basic UI elements are present. Opening of folders works, but error-handling is not complete. Images can be displayed and bounding-boxes can be created and both are dynamically resized based on the window size. The bounding-box stays correctly anchored to its position on the image if the window is resized. Iterating through the image files using the next/previous buttons works. Brightness slider is implemented, zoom is not. Bounding-box data is not yet stored anywhere.