Add settings sidepane and boundingbox saving functionality
Zoom-slider and brightness-slider and boundingbox color-picker are now inside a settings side pane on the
right side of the window. Zoom-slider is not yet synchronized with CTRL-Scroll zooming. Saving of the drawn
boundingboxes to a csv/txt file is implemented. Drawing of SelectionRectangles now (correctly) works only inside
the current image bounds (even if the image contains transparent parts), resizing using the ResizeHandles
needs to be tuned to clamp to these bounds too.